Removed obsolete SearchModel class
