+++ test/DynamoCoreTests/SearchViewModelTests.cs
@@ -54,20 +54,20 @@ namespace Dynamo.Tests
         public void ShortenCategoryNameTests()
         {
             var categoryName = "";
-            var result = SearchModel.ShortenCategoryName(categoryName);
+            var result = Nodes.Utilities.ShortenCategoryName(categoryName);
             Assert.AreEqual(string.Empty, result);
 
             categoryName = null;
-            result = SearchModel.ShortenCategoryName(categoryName);
+            result = Nodes.Utilities.ShortenCategoryName(categoryName);
             Assert.AreEqual(string.Empty, result);
 
             categoryName = "Category1";
-            result = SearchModel.ShortenCategoryName(categoryName);
+            result = Nodes.Utilities.ShortenCategoryName(categoryName);
             Assert.AreEqual("Category1", result);
 
             categoryName = "Cat1 Cat" + Configurations.CategoryDelimiter + "Cat2 Cat" +
                                     Configurations.CategoryDelimiter + "Cat3";
-            result = SearchModel.ShortenCategoryName(categoryName);
+            result = Nodes.Utilities.ShortenCategoryName(categoryName);
             Assert.AreEqual("Cat1 Cat " + Configurations.ShortenedCategoryDelimiter + " Cat2 Cat " +
                                       Configurations.ShortenedCategoryDelimiter + " Cat3", result);
 
@@ -77,7 +77,7 @@ namespace Dynamo.Tests
                            "TenSymbol" + Configurations.CategoryDelimiter +
                            "TenSymbol" + Configurations.CategoryDelimiter +
                            "MoreSymbols";
-            result = SearchModel.ShortenCategoryName(categoryName);
+            result = Nodes.Utilities.ShortenCategoryName(categoryName);
             Assert.AreEqual("TenSymbol " + Configurations.ShortenedCategoryDelimiter +
                            " ... " + Configurations.ShortenedCategoryDelimiter +
                            " TenSymbol " + Configurations.ShortenedCategoryDelimiter +
